Rename _i_observable to i_observable
