+++ traits/ctrait.py
@@ -20,7 +20,7 @@ import inspect
 
 from . import ctraits
 from .constants import ComparisonMode, DefaultValue, default_value_map
-from .observation._i_observable import IObservable
+from .observation.i_observable import IObservable
 from .trait_base import SequenceTypes, Undefined
 from .trait_dict_object import TraitDictObject
 from .trait_list_object import TraitListObject
