+++ traits/trait_dict_object.py
@@ -11,7 +11,7 @@
 import copy
 from weakref import ref
 
-from traits.observation._i_observable import IObservable
+from traits.observation.i_observable import IObservable
 from traits.trait_base import Undefined, _validate_everything
 from traits.trait_errors import TraitError
