+++ traits/trait_set_object.py
@@ -13,7 +13,7 @@ import copyreg
 from itertools import chain
 from weakref import ref
 
-from traits.observation._i_observable import IObservable
+from traits.observation.i_observable import IObservable
 from traits.trait_base import _validate_everything
 from traits.trait_errors import TraitError
